Extract course sync backfill helper
